requirements.txt
polars>=0.20
fsspec
s3fs
python-dotenv
pandas
pyarrow
